session2/1.stockresearch/pyproject.toml
[project]
name = "stockresearch"
version = "0.1.0"
description = "stockResearch using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.12,<3.13"
dependencies = [
    "boto3",
    "crewai[anthropic,tools]>=0.130.0",
    "langfuse>=3.5.0",
    "openinference-instrumentation-crewai>=0.1.13",
    "openinference-instrumentation-litellm",
    "opentelemetry-exporter-otlp-proto-http",
    "openlit",
    "crewai_tools",
    "litellm",
    "bedrock-agentcore-starter-toolkit>=0.3.5",
    "bedrock-agentcore>=1.6.2",
    "opentelemetry-api",
    "opentelemetry-exporter-otlp",
    "opentelemetry-sdk",
]

[project.scripts]
stockresearch = "stockresearch.main:run"
run_crew = "stockresearch.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
